+++ test/bloomfilter/bloomfilter_test.go
@@ -14,7 +14,7 @@ package bloomfilter
 import (
 	"testing"
 
-	bf "github.com/justincpresley/ndn-sync/internal/bloomfilter"
+	bf "github.com/justincpresley/ndn-sync/util/bloomfilter"
 	assert "github.com/stretchr/testify/assert"
 )
 
@@ -25,4 +25,21 @@ func TestBasic(t *testing.T) {
 
 	assert.True(t, b.Check(d1))
 	assert.False(t, b.Check(d2))
+}
+
+func TestEncode(t *testing.T) {
+	b := bf.NewFilter(3, 100)
+	b.Add([]byte("TEST_ENTRY1"))
+	b.Add([]byte("ANOTHER_ENTRY2"))
+
+	buf, err := b.Bytes()
+	assert.Nil(t, err)
+	assert.Equal(t, []byte{3, 4, 0, 0, 0, 0, 0, 64, 0, 0, 2, 0, 65, 68, 0}, buf)
+}
+
+func TestDecode(t *testing.T) {
+	b, err := bf.ParseFilter([]byte{3, 4, 0, 0, 0, 0, 0, 64, 0, 0, 2, 0, 65, 68, 0})
+	assert.Nil(t, err)
+	assert.True(t, b.Check([]byte("TEST_ENTRY1")))
+	assert.True(t, b.Check([]byte("ANOTHER_ENTRY2")))
 }